refactor(web): move mr_rework opt-in to Run defaults + fixture/copy/a11y fixes (PRD #700 M6 review)
Address M6 review findings: move the "Auto-rework MR review comments" per-user
opt-in out of the Account & tokens tab into Run defaults, beside the autopilot/
judge/CI-autofix run-behavior toggles (the documented IA split). Soften the
helper copy so resolve is described as forge-dependent (GitLab/GitHub; Forgejo is
reply-only). Make the mock run-mr-rework fixture issue-less (issue_iid=null,
issue_web_url=null) to match the shipped backend (CreateAutoMRReworkRun leaves
issue_iid NULL), so it renders the "MR rework" run-kind chip, and add a JSX render
test exercising that path. Link the helper copy to the checkbox via aria-describedby.
